requirements.txt
moviepy~=1.0.3
pillow~=11.3.0
elevenlabs~=2.16.0
protobuf~=6.33.0
PyYAML~=6.0.3
pydantic~=2.11.9
python-dotenv~=1.1.1
google-api-python-client~=2.186.0